PR feedback and bug fix for type conversion
